requirements.txt
aiohttp==3.7.4.post0
annotated-types==0.7.0
anyio==4.4.0
async-timeout==3.0.1
attrs==21.4.0
bcrypt==4.0.1
beautifulsoup4==4.11.1
blinker==1.6.2
certifi==2022.6.15
cffi==1.15.1
chardet==4.0.0
charset-normalizer==2.1.0
click==8.1.3
colorama==0.4.5
discord.py==1.7.3
dnspython==2.4.2
docopt==0.6.2
email-validator==2.0.0.post2
exceptiongroup==1.2.2
fastapi==0.112.1
Flask==2.3.3
Flask-Bcrypt==1.0.1
Flask-Cors==5.0.0
Flask-Login==0.6.2
Flask-Mail==0.9.1
Flask-SQLAlchemy==3.0.5
Flask-WTF==1.1.1
gevent==21.12.0
greenlet==1.1.2
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.3
itsdangerous==2.1.2
Jinja2==3.1.2
keyboard==0.13.5
MarkupSafe==2.1.1
MouseInfo==0.1.3
multidict==6.0.2
mysql==0.0.3
mysql-connector-python==8.0.29
mysqlclient==2.1.1
notion-client==2.3.0
Pillow==10.0.0
pipreqs==0.4.11
protobuf==4.21.2
PyAutoGUI==0.9.53
pycparser==2.21
pydantic==2.8.2
pydantic_core==2.20.1
pygame==2.6.0
PyGetWindow==0.0.9
PyMsgBox==1.0.9
pyperclip==1.8.2
PyRect==0.2.0
PyScreeze==0.1.28
python-dotenv==1.0.0
pytweening==1.0.4
requests==2.28.1
sniffio==1.3.1
soupsieve==2.3.2.post1
SQLAlchemy==2.0.20
starlette==0.38.2
typing_extensions==4.12.2
ultraimport==0.0.7
urllib3==1.26.9
uvicorn==0.30.6
websocket==0.2.1
websocket-client==1.3.3
websockets==10.3
Werkzeug==2.3.7
WTForms==3.0.1
yarg==0.1.9
yarl==1.7.2
zope.event==4.5.0
zope.interface==5.4.0
